Character Inheritance
Removed old chracater classes added new base class and a diesel and
cyber hero class that inherits from it, made blueprint for each hero as
well.
