+++ Binaries/Win64/FusionpunksEditor.target
@@ -2,7 +2,7 @@
 	"TargetName" : "FusionpunksEditor",
 	"Platform" : "Win64",
 	"Configuration" : "Development",
-	"BuildId" : "21abb897-0211-4b52-997e-efd5b335a37e",
+	"BuildId" : "b5d76bfd-de08-4c5f-b19a-7de281e9875d",
 	"Version" : 
 	{
 		"MajorVersion" : 4,
