+++ Binaries/Win64/UE4Editor.modules
@@ -1,8 +1,8 @@
 {
 	"Changelist" : 3039270,
-	"BuildId" : "21abb897-0211-4b52-997e-efd5b335a37e",
+	"BuildId" : "b5d76bfd-de08-4c5f-b19a-7de281e9875d",
 	"Modules" : 
 	{
-		"Fusionpunks" : "UE4Editor-Fusionpunks.dll"
+		"Fusionpunks" : "UE4Editor-Fusionpunks-796.dll"
 	}
 }
