+++ Source/Fusionpunks/FusionpunksGameMode.cpp
@@ -3,12 +3,12 @@
 #include "Fusionpunks.h"
 #include "FusionpunksGameMode.h"
 #include "PlayerHUD.h"
-#include "FusionpunksCharacter.h"
+#include "DieselHeroCharacter.h"
 
 AFusionpunksGameMode::AFusionpunksGameMode()
 {
 	// set default pawn class to our Blueprinted character
-	static ConstructorHelpers::FClassFinder<APawn> PlayerPawnBPClass(TEXT("/Game/ThirdPersonCPP/Blueprints/ThirdPersonCharacter"));
+	static ConstructorHelpers::FClassFinder<APawn> PlayerPawnBPClass(TEXT("Blueprint'/Game/Blueprints/DieselCharacter.DieselCharacter'"));
 	if (PlayerPawnBPClass.Class != NULL)
 	{
 		DefaultPawnClass = PlayerPawnBPClass.Class;
