+++ Source/Fusionpunks/HeroBase.h
@@ -1,0 +1,97 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "GameFramework/Character.h"
+#include "HeroBase.generated.h"
+
+UCLASS()
+class FUSIONPUNKS_API AHeroBase : public ACharacter
+{
+	GENERATED_BODY()
+
+	/** Camera boom positioning the camera behind the character */
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera, meta = (AllowPrivateAccess = "true"))
+		class USpringArmComponent* CameraBoom;
+
+	/** Follow camera */
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera, meta = (AllowPrivateAccess = "true"))
+		class UCameraComponent* FollowCamera;
+public:
+	AHeroBase();
+
+	virtual void BeginPlay() override;
+
+	virtual void Tick(float DeltaTime) override;
+
+	// APawn interface
+	virtual void SetupPlayerInputComponent(class UInputComponent* InputComponent) override;
+	// End of APawn interface
+	/** Base turn rate, in deg/sec. Other scaling may affect final turn rate. */
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera)
+		float BaseTurnRate;
+
+	/** Base look up/down rate, in deg/sec. Other scaling may affect final rate. */
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera)
+		float BaseLookUpRate;
+
+protected:
+	/** Called for forwards/backward input */
+	void MoveForward(float Value);
+	/** Called for side to side input */
+	void MoveRight(float Value);
+	/**
+	* Called via input to turn at a given rate.
+	* @param Rate	This is a normalized rate, i.e. 1.0 means 100% of desired turn rate
+	*/
+	void TurnAtRate(float Rate);
+	/**
+	* Called via input to turn look up/down at a given rate.
+	* @param Rate	This is a normalized rate, i.e. 1.0 means 100% of desired turn rate
+	*/
+	void LookUpAtRate(float Rate);
+
+
+	
+
+public:
+	/** Returns CameraBoom subobject **/
+	FORCEINLINE class USpringArmComponent* GetCameraBoom() const { return CameraBoom; }
+	/** Returns FollowCamera subobject **/
+	FORCEINLINE class UCameraComponent* GetFollowCamera() const { return FollowCamera; }
+
+public:
+	UFUNCTION(BlueprintCallable, Category = HeroFunctions)
+		float GetPlayerHealthPercentage();
+	UFUNCTION(BlueprintCallable, Category = HeroFunctions)
+		float GetCurrentHealth();
+	UFUNCTION(BlueprintCallable, Category = HeroFunctions)
+		float GetMaxHealth();
+
+protected:
+	UPROPERTY(EditAnywhere, Category = Stats) //Make private for final game
+	class ACreepCamp* CurrentCreepCamp;
+	void StartAttack();
+	void Attack(AActor* enemy);
+	void AdjustCameraZoom(float Value);
+	float currentHealth;
+	float maxHealth;
+	
+public:
+	
+	UFUNCTION(BlueprintCallable, Category = UIFunctions)
+		void ShowCampProgress(ACreepCamp* CurrentCamp);
+
+	UFUNCTION(BlueprintCallable, Category = UIFunctions)
+		void HideCampProgress();
+private:
+	//function for Trigger Events
+	UFUNCTION()
+		void OnOverlapBegin(class UPrimitiveComponent* ThisComp, class AActor* OtherActor, class UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult &SweepResult);
+
+	//function for Trigger Exit Events
+	UFUNCTION()
+		void OnOverlapEnd(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex);
+
+	
+};
