+++ Source/Fusionpunks/PlayerHealthBarWidget.cpp
@@ -1,21 +1,21 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #include "Fusionpunks.h"
-#include "FusionpunksCharacter.h"
+#include "HeroBase.h"
 #include "PlayerHealthBarWidget.h"
 
 #define LOCTEXT_NAMESPACE "MyNamespace"
 
 float UPlayerHealthBarWidget::GetPlayerHealthPercentage() const
 {
-	AFusionpunksCharacter* MyCharacter = Cast<AFusionpunksCharacter>(GetOwningPlayerPawn());
+	AHeroBase* MyCharacter = Cast<AHeroBase>(GetOwningPlayerPawn());
 
 	return MyCharacter->GetCurrentHealth() / MyCharacter->GetMaxHealth() * 100;
 }
 
 FText UPlayerHealthBarWidget::GetPlayerHealthText() const
 {
-	AFusionpunksCharacter* MyCharacter = Cast<AFusionpunksCharacter>(GetOwningPlayerPawn());
+	AHeroBase* MyCharacter = Cast<AHeroBase>(GetOwningPlayerPawn());
 
 	return FText::Format(LOCTEXT("HealthFormat", "Health: {0} / {1}"),
 		FText::AsNumber(MyCharacter->GetCurrentHealth()),
