+++ Source/Fusionpunks/Projectile.cpp
@@ -2,7 +2,7 @@
 
 #include "Fusionpunks.h"
 #include "Projectile.h"
-#include "MobaCharacter.h"
+#include "CyberHeroCharacter.h"
 
 
 // Sets default values
